move over everything first
