+++ components/camera/transformpipeline/classifier_test.go
@@ -4,6 +4,7 @@ package transformpipeline
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"testing"
 
@@ -108,6 +109,7 @@ func TestClassifierSource(t *testing.T) {
 	r, err := buildRobotWithClassifier(logger)
 	test.That(t, err, test.ShouldBeNil)
 
+	fmt.Print(r.ResourceNames())
 	defer func() {
 		test.That(t, r.Close(context.Background()), test.ShouldBeNil)
 	}()
