+++ components/camera/transformpipeline/mods_test.go
@@ -10,7 +10,7 @@ import (
 	"go.viam.com/utils/artifact"
 
 	"go.viam.com/rdk/components/camera"
-	"go.viam.com/rdk/components/camera/videosource"
+	"go.viam.com/rdk/components/camera/fake"
 	"go.viam.com/rdk/gostream"
 	"go.viam.com/rdk/rimage"
 	"go.viam.com/rdk/utils"
@@ -30,7 +30,7 @@ func TestCrop(t *testing.T) {
 	test.That(t, err, test.ShouldBeNil)
 
 	// test depth source
-	source := gostream.NewVideoSource(&videosource.StaticSource{DepthImg: dm}, prop.Video{})
+	source := gostream.NewVideoSource(&fake.StaticSource{DepthImg: dm}, prop.Video{})
 	out, _, err := camera.ReadImage(context.Background(), source)
 	test.That(t, err, test.ShouldBeNil)
 	test.That(t, out.Bounds().Dx(), test.ShouldEqual, 128)
@@ -48,7 +48,7 @@ func TestCrop(t *testing.T) {
 	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 
 	// test color source
-	source = gostream.NewVideoSource(&videosource.StaticSource{ColorImg: img}, prop.Video{})
+	source = gostream.NewVideoSource(&fake.StaticSource{ColorImg: img}, prop.Video{})
 	out, _, err = camera.ReadImage(context.Background(), source)
 	test.That(t, err, test.ShouldBeNil)
 	test.That(t, out.Bounds().Dx(), test.ShouldEqual, 128)
@@ -107,7 +107,7 @@ func TestResizeColor(t *testing.T) {
 		"height_px": 20,
 		"width_px":  30,
 	}
-	source := gostream.NewVideoSource(&videosource.StaticSource{ColorImg: img}, prop.Video{})
+	source := gostream.NewVideoSource(&fake.StaticSource{ColorImg: img}, prop.Video{})
 	out, _, err := camera.ReadImage(context.Background(), source)
 	test.That(t, err, test.ShouldBeNil)
 	test.That(t, out.Bounds().Dx(), test.ShouldEqual, 128)
@@ -133,7 +133,7 @@ func TestResizeDepth(t *testing.T) {
 		"height_px": 40,
 		"width_px":  60,
 	}
-	source := gostream.NewVideoSource(&videosource.StaticSource{DepthImg: img}, prop.Video{})
+	source := gostream.NewVideoSource(&fake.StaticSource{DepthImg: img}, prop.Video{})
 	out, _, err := camera.ReadImage(context.Background(), source)
 	test.That(t, err, test.ShouldBeNil)
 	test.That(t, out.Bounds().Dx(), test.ShouldEqual, 128)
@@ -154,7 +154,7 @@ func TestRotateColorSource(t *testing.T) {
 	img, err := rimage.NewImageFromFile(artifact.MustPath("rimage/board1_small.png"))
 	test.That(t, err, test.ShouldBeNil)
 
-	source := gostream.NewVideoSource(&videosource.StaticSource{ColorImg: img}, prop.Video{})
+	source := gostream.NewVideoSource(&fake.StaticSource{ColorImg: img}, prop.Video{})
 	am := utils.AttributeMap{
 		"angle_degs": 180,
 	}
@@ -207,7 +207,7 @@ func TestRotateColorSource(t *testing.T) {
 	test.That(t, rs.Close(context.Background()), test.ShouldBeNil)
 	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 
-	source = gostream.NewVideoSource(&videosource.StaticSource{ColorImg: img}, prop.Video{})
+	source = gostream.NewVideoSource(&fake.StaticSource{ColorImg: img}, prop.Video{})
 	am = utils.AttributeMap{
 		"angle_degs": 90,
 	}
@@ -238,7 +238,7 @@ func TestRotateColorSource(t *testing.T) {
 	test.That(t, rs.Close(context.Background()), test.ShouldBeNil)
 	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 
-	source = gostream.NewVideoSource(&videosource.StaticSource{ColorImg: img}, prop.Video{})
+	source = gostream.NewVideoSource(&fake.StaticSource{ColorImg: img}, prop.Video{})
 	am = utils.AttributeMap{
 		"angle_degs": -90,
 	}
@@ -269,7 +269,7 @@ func TestRotateColorSource(t *testing.T) {
 	test.That(t, rs.Close(context.Background()), test.ShouldBeNil)
 	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 
-	source = gostream.NewVideoSource(&videosource.StaticSource{ColorImg: img}, prop.Video{})
+	source = gostream.NewVideoSource(&fake.StaticSource{ColorImg: img}, prop.Video{})
 	am = utils.AttributeMap{
 		"angle_degs": 270,
 	}
@@ -300,7 +300,7 @@ func TestRotateColorSource(t *testing.T) {
 	test.That(t, rs.Close(context.Background()), test.ShouldBeNil)
 	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 
-	source = gostream.NewVideoSource(&videosource.StaticSource{ColorImg: img}, prop.Video{})
+	source = gostream.NewVideoSource(&fake.StaticSource{ColorImg: img}, prop.Video{})
 	am = utils.AttributeMap{
 		"angle_degs": 0, // no-op
 	}
@@ -336,7 +336,7 @@ func TestRotateDepthSource(t *testing.T) {
 		context.Background(), artifact.MustPath("rimage/board1_gray_small.png"))
 	test.That(t, err, test.ShouldBeNil)
 
-	source := gostream.NewVideoSource(&videosource.StaticSource{DepthImg: pc}, prop.Video{})
+	source := gostream.NewVideoSource(&fake.StaticSource{DepthImg: pc}, prop.Video{})
 	am := utils.AttributeMap{
 		"angle_degs": 180,
 	}
@@ -388,7 +388,7 @@ func TestRotateDepthSource(t *testing.T) {
 	test.That(t, rs.Close(context.Background()), test.ShouldBeNil)
 	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 
-	source = gostream.NewVideoSource(&videosource.StaticSource{DepthImg: pc}, prop.Video{})
+	source = gostream.NewVideoSource(&fake.StaticSource{DepthImg: pc}, prop.Video{})
 	am = utils.AttributeMap{
 		"angle_degs": 90,
 	}
@@ -419,7 +419,7 @@ func TestRotateDepthSource(t *testing.T) {
 	test.That(t, rs.Close(context.Background()), test.ShouldBeNil)
 	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 
-	source = gostream.NewVideoSource(&videosource.StaticSource{DepthImg: pc}, prop.Video{})
+	source = gostream.NewVideoSource(&fake.StaticSource{DepthImg: pc}, prop.Video{})
 	am = utils.AttributeMap{
 		"angle_degs": -90,
 	}
@@ -450,7 +450,7 @@ func TestRotateDepthSource(t *testing.T) {
 	test.That(t, rs.Close(context.Background()), test.ShouldBeNil)
 	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 
-	source = gostream.NewVideoSource(&videosource.StaticSource{DepthImg: pc}, prop.Video{})
+	source = gostream.NewVideoSource(&fake.StaticSource{DepthImg: pc}, prop.Video{})
 	am = utils.AttributeMap{
 		"angle_degs": 270,
 	}
@@ -481,7 +481,7 @@ func TestRotateDepthSource(t *testing.T) {
 	test.That(t, rs.Close(context.Background()), test.ShouldBeNil)
 	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 
-	source = gostream.NewVideoSource(&videosource.StaticSource{DepthImg: pc}, prop.Video{})
+	source = gostream.NewVideoSource(&fake.StaticSource{DepthImg: pc}, prop.Video{})
 	am = utils.AttributeMap{
 		"angle_degs": 0, // no-op
 	}
@@ -516,7 +516,7 @@ func BenchmarkColorRotate(b *testing.B) {
 	img, err := rimage.NewImageFromFile(artifact.MustPath("rimage/board1.png"))
 	test.That(b, err, test.ShouldBeNil)
 
-	source := gostream.NewVideoSource(&videosource.StaticSource{ColorImg: img}, prop.Video{})
+	source := gostream.NewVideoSource(&fake.StaticSource{ColorImg: img}, prop.Video{})
 	src, err := camera.WrapVideoSourceWithProjector(context.Background(), source, nil, camera.ColorStream)
 	test.That(b, err, test.ShouldBeNil)
 	am := utils.AttributeMap{
@@ -540,7 +540,7 @@ func BenchmarkDepthRotate(b *testing.B) {
 		context.Background(), artifact.MustPath("rimage/board1.dat.gz"))
 	test.That(b, err, test.ShouldBeNil)
 
-	source := gostream.NewVideoSource(&videosource.StaticSource{DepthImg: img}, prop.Video{})
+	source := gostream.NewVideoSource(&fake.StaticSource{DepthImg: img}, prop.Video{})
 	src, err := camera.WrapVideoSourceWithProjector(context.Background(), source, nil, camera.DepthStream)
 	test.That(b, err, test.ShouldBeNil)
 	am := utils.AttributeMap{
